Modularize features and make the list interactive
