+++ src/index.html
@@ -10,7 +10,7 @@
 </head>
 <body>
   <div class="container">
-    <h1>Today's To Do <i class="fa-solid fa-arrows-rotate"></i></h1>
+    <h1>Today's To Do <i class="fa-solid fa-arrows-rotate refresh"></i></h1>
     <form>
       <input type="text" class="addToDo" required placeholder="Add to your list...">
       <i class="fa-solid fa-right-to-bracket addBtn"></i>
